Moved collectorsrc into package with in project src.
