+++ src/main/java/collectors/CreditSupply.java
@@ -1,15 +1,16 @@
-package housing;
+package collectors;
 
 import java.io.FileNotFoundException;
 import java.io.PrintWriter;
 import java.io.UnsupportedEncodingException;
 
+import housing.*;
 import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
 
 public class CreditSupply extends CollectorBase {
 	private static final long serialVersionUID = 1630707025974306844L;
 
-	private Config	config = Model.config;	// Passes the Model's configuration parameters object to a private field
+	private Config config = Model.config;	// Passes the Model's configuration parameters object to a private field
 
 	public CreditSupply() {
 		mortgageCounter = 0;
