+++ src/main/java/housing/CentralBank.java
@@ -26,7 +26,7 @@ public class CentralBank implements Serializable {
 	 * This method implements the policy strategy of the Central Bank.
 	 * @param coreIndicators The current value of the core indicators
 	 */
-	public void step(CoreIndicators coreIndicators) {
+	public void step(collectors.CoreIndicators coreIndicators) {
 		/** Use this method to express the policy strategy of the central bank by
 		 * setting the value of the various limits in response to the current
 		 * value of the core indicators.
